wsl: consolidate and simplify WSL2 distro cleanup
Apply review feedback from PR #479.

- Move the distro cleanup helpers from package service into pkg/util/wsl,
  so all wsl.exe management lives in one package.
- Capture wsl.exe's output in the wrapped error and set WSL_UTF8, so a
  failed terminate or unregister reports a legible reason.
- Drop the terminate before the orphan-heal unregister: a missing root
  disk means the distro was stopped, so it holds no kernel state for
  --unregister to deadlock on.
- Inline unregisterWSL2Distro, a pure rename whose two callers handle the
  error differently; keep terminateWSL2Distro, which consolidates the
  best-effort debug logging its callers share.
