+++ rdd/pkg/service/cmd/service.go
@@ -566,64 +566,9 @@ func Delete(ctx context.Context, timeout time.Duration) error {
 	// directory deletes the distro's ext4.vhdx but orphans the `wsl --list`
 	// registration, which the next `rdd svc create` would boot against a
 	// missing disk.
-	unregisterInstanceDistros(ctx)
+	wsl.UnregisterInstanceDistros(ctx, instance.LimaHome())
 	_ = os.RemoveAll(instance.ShortDir())
 	return os.RemoveAll(instance.Dir())
-}
-
-// unregisterInstanceDistros unregisters the WSL2 distro backing each Lima
-// instance, clearing the `wsl --list` registration before its directory is
-// removed. It terminates each distro first, because `wsl --unregister` can
-// deadlock on a distro that still holds kernel state. Both calls are best-effort
-// and time-bounded, since wsl.exe can hang when the WSL subsystem is degraded.
-// No-op on non-Windows.
-//
-// TODO: This reimplements cleanup the lima controller already owns
-// (unregisterWSL2Distro), because `svc delete` stops the daemon before
-// reaching here. Route it through the controller once the stop/delete hook
-// (rancher-desktop-2#319) exists.
-func unregisterInstanceDistros(ctx context.Context) {
-	if runtime.GOOS != "windows" {
-		return
-	}
-	for _, distroName := range instanceDistroNames(instance.LimaHome()) {
-		// Terminating is a best-effort precursor — the distro is usually already
-		// stopped — so log its failure at debug.
-		if err := wsl.Terminate(ctx, distroName); err != nil {
-			logrus.WithError(err).WithField("distro", distroName).Debug("wsl --terminate failed during delete")
-		}
-
-		// A failed unregister is the failure that matters: it leaves behind the
-		// stale registration this cleanup exists to prevent, so log it at warn.
-		// (A distro that was never imported also fails here, harmlessly.)
-		if err := wsl.Unregister(ctx, distroName); err != nil {
-			logrus.WithError(err).WithField("distro", distroName).Warn("Failed to unregister WSL2 distro; a stale registration may remain")
-		}
-	}
-}
-
-// instanceDistroNames returns the WSL2 distro name ("lima-<instance>") for each
-// Lima instance directory under limaHome. It skips Lima's reserved bookkeeping
-// directories (those prefixed with "_" or "."), which are never instances, the
-// same way Lima's own store.Instances does. It returns nil when limaHome does
-// not exist (no instances were ever created).
-func instanceDistroNames(limaHome string) []string {
-	entries, err := os.ReadDir(limaHome)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			logrus.WithError(err).Warn("Failed to read Lima instance directory for WSL2 cleanup")
-		}
-		return nil
-	}
-	var names []string
-	for _, entry := range entries {
-		name := entry.Name()
-		if !entry.IsDir() || strings.HasPrefix(name, "_") || strings.HasPrefix(name, ".") {
-			continue
-		}
-		names = append(names, wsl.DistroName(name))
-	}
-	return names
 }
 
 // preserveAllInstanceLogs moves .log files from each Lima instance directory
